#182 - Add video playback controls to video helper and move it from Hero video, fix playback button dimensions
