+++ blocks/v2-hero/v2-hero.css
@@ -146,37 +146,6 @@
   width: var(--scroll-dot-width);
 }
 
-.v2-hero .icon-play-video {
-  display: none;
-  justify-content: center;
-}
-
-.v2-hero .icon-pause-video {
-  display: flex;
-  justify-content: center;
-}
-
-.v2-hero .v2-hero__playback-button {
-  position: absolute;
-  bottom: 20px;
-  right: 20px;
-  background: transparent;
-  width: 44px;
-  height: 44px;
-  border-radius: 16px;
-  border: 0;
-  padding: 0;
-}
-
-.v2-hero .v2-hero__playback-button:hover svg circle {
-  fill: var(--c-grey-50);
-}
-
-.v2-hero .v2-hero__playback-button .icon svg {
-  width: 26px;
-  height: 26px;
-}
-
 @media (min-width: 744px) {
   .v2-hero {
     --v2-hero-padding: 50px;
