+++ blocks/v2-video/v2-video.js
@@ -1,4 +1,4 @@
-import { createVideo } from '../../scripts/video-helper.js';
+import { addPlayIcon, createVideo } from '../../scripts/video-helper.js';
 
 const onHoverOrScroll = (element, handler) => {
   let isInViewport = false;
@@ -67,4 +67,11 @@ export default async function decorate(block) {
 
     block.classList[action](`${blockClass}--full-width`);
   });
+
+  const playbackControls = document.querySelectorAll('.v2-hero__video .v2-video__playback-button, .v2-video__video .v2-video__playback-button');
+  playbackControls.forEach((control) => {
+    const parentElement = control.parentElement.parentElement;
+    const targetElement = parentElement.querySelector('video + div')
+    targetElement.append(control);
+  })
 }
